// src/types/Team.ts
/**
 * @author: JP Lew (jp@cto.ai)
 * @date: Monday, 15th April 2019 2:29:03 pm
 * @lastModifiedBy: JP Lew (jp@cto.ai)
 * @lastModifiedTime: Tuesday, 16th April 2019 8:19:04 am
 * @copyright (c) 2019 CTO.ai
 */

export default interface Team {
  id: string
  name: string
  createdBy: string
  createdAt: string
  updatedAt: string
}